% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_rpart}
\alias{convert_rpart}
\title{Convert rpart to d3.js hierarchy}
\usage{
convert_rpart(rpart_data = NULL)
}
\arguments{
\item{rpart_data}{\code{rpart} object to be converted}
}
\description{
This thing is not even close to being done.
}

