% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R, R/sankeytree.R
\docType{package}
\name{sankeytree}
\alias{sankeytree}
\alias{sankeytree-package}
\title{Sankey Diagrams as Collapsible Trees}
\usage{
sankeytree(data = NULL, maxLabelLength = NULL, nodeHeight = NULL,
  width = NULL, height = NULL)
}
\description{
Combining Sankey diagrams with collapsible trees and adding some new interactivity might help us analyze, instruct, and decide.

<Add Description>
}

